separation of triangulation and constraint util - further code refactoring
