+++ Runtime/CDT/CDT.ConstraintUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9297d0a6e1c40384db1eb8b5d0f9244c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
